+++ framework.orm/src/main/java/work/myfavs/framework/orm/meta/dialect/SqlServer2012Dialect.java
@@ -18,21 +18,21 @@ public class SqlServer2012Dialect
   }
 
   @Override
-  public Sql selectTop(long currentPage, long pageSize, String sql, List<Object> params) {
+  public Sql selectTop(int currentPage, int pageSize, String sql, List<Object> params) {
 
     Sql querySql = new Sql(sql, params);
-    if (pageSize == -1L) {
+    if (pageSize == -1) {
       return querySql;
     }
 
-    if (currentPage == 1L && pageSize == 1L) {
+    if (currentPage == 1 && pageSize == 1) {
       //如果sql本身只返回一个结果
       if (selectSinglePattern.matcher(sql).find()) {
         return querySql;
       }
     }
 
-    long offset = pageSize * (currentPage - 1L);
+    int offset = pageSize * (currentPage - 1);
     return querySql.append("OFFSET ? ROWS FETCH NEXT ? ROWS ONLY", offset, pageSize);
   }
